% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CCEI.R
\name{calculate_CCEI}
\alias{calculate_CCEI}
\title{Calculate_CCEI}
\usage{
calculate_CCEI(df, step, ...)
}
\arguments{
\item{df}{a dataframe}

\item{step}{}
}
\value{
the CCEI score
}
\description{
\code{calculate_CCEI}
}
\examples{
set.seed(1234)
df <- gather_data(simGarp(10, 2, afriat.par = 0.5))
calculate_CCEI(df, step = 0.001, x1, x2, p1, p2)
}
