% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSarpExtra.R
\name{checkSarpExtra}
\alias{checkSarpExtra}
\title{Check Sarp Extra}
\usage{
checkSarpExtra(df, ...)
}
\arguments{
\item{df}{a dataframe containing columns with quantities and prices of goods}

\item{...}{the columns within \code{df} that contain quantities and prices. 
The first half of the columns provided are assumed to be quantities and the
second half of the columns provided are be assumed to be prices.}
}
\value{

}
\description{
\code{checkSarpExtra} is a wrapper function for the
\code{\link[revealedPrefs]{checkSarp}} function from the \pkg{revealedPrefs} package.
}
\examples{
experiment_data<- tibble(subject = c(1, 1, 2, 2),
                              x1 = c(1:4),
                              x2 = c(5:8),
                              p1 = c(5:8),
                              p2 = c(5:8))
 
experiment_data \%>\%
 checkSarpExtra(x1, x2, p1, p2)
}
