% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_data.R
\name{gather_data}
\alias{gather_data}
\title{Gather Data}
\usage{
gather_data(matrix_list)
}
\arguments{
\item{matrix_list}{a list containg two matrices:
\describe{
  \item{x}{a matrix containing quantities}
  \item{p}{a matrix containing prices}
}}
}
\value{
a dataframe where the rows represent an individual's choice and the columns
represent the quantities and prices of a good
}
\description{
\code{gather_data} is written to work with functions from 
the \pkg{revealedPrefs} package.

\code{\link[revealedPrefs]{checkWarp}} and similar "check" functions
take as input two \code{n} by \code{i} matrices, one for goods and one for prices. 
\code{gather_data} takes a list with two matrices and returns a dataframe.
}
\examples{
simGarp(2,4) \%>\%
gather_data()
}
