% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot Data}
\usage{
plot_data(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
\code{plot_data} does something cool.
}
\examples{
simGarp(4, 2) \%>\%
spread_data() \%>\%
plot_data()
}
