% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_data.R
\name{spread_data}
\alias{spread_data}
\title{Spread Data}
\usage{
spread_data(df, ...)
}
\arguments{
\item{df}{a dataframe containing columns with quantities and prices of goods}

\item{...}{the columns within \code{df} that contain prices}
}
\value{
a list containg two matrices:
\describe{
  \item{x}{a matrix containing quantities}
  \item{p}{a matrix containing prices}
}
}
\description{
\code{spread_data} is written to work with functions from 
the \pkg{revealedPrefs} package.

\code{\link[revealedPrefs]{checkWarp}} and similar "check" functions
take as input two \code{n} by \code{i} matrices, one for goods and one for prices. 
\code{spread_data} takes a dataframe and returns a list with two matrices, which
can then be passed into \code{\link[revealedPrefs]{checkWarp}} 
and similar functions.
}
\examples{
# simGarp is a function from revealedPrefs that creates a 
# a random data set that is returned as a list containing two
# matrices
simGarp(4, 2) \%>\%
spread_data()
}
