% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CCEI.R
\name{calculate_CCEI}
\alias{calculate_CCEI}
\title{Calculate_CCEI}
\usage{
calculate_CCEI(df, step = 0.01)
}
\arguments{
\item{df}{a dataframe}

\item{step}{the step does something}
}
\value{
a integer
}
\description{
\code{calculate_CCEI} does something cool.
}
\examples{
calculate_CCEI(simGarp(10, 2, afriat.par = 0.5), step = 0.001)
}
