% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_data.R
\name{spread_data}
\alias{spread_data}
\title{Spread Data}
\usage{
spread_data(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a tibble
}
\description{
\code{spread_data} does something cool.
}
\examples{
# simGarp is a function from revealedPrefs that creates a 
# a random data set that is returned as a list containing two
# matrices
simGarp(4, 2) \%>\%
spread_data()
}
